+++ User/app/src/main/java/com/example/jub_jub_android/ui/adapter/viewpager/MyRentList_ViewPagerAdapter.kt
@@ -1,19 +1,20 @@
-package com.example.jub_jub_android.ui.adapter
+package com.example.jub_jub_android.ui.adapter.viewpager
 
 import android.content.Context
-import androidx.viewpager.widget.PagerAdapter
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.viewpager.widget.PagerAdapter
 import com.example.jub_jub_android.R
-import com.example.jub_jub_android.entity.singleton.ManageItemListManager
-import com.example.jub_jub_android.entity.singleton.ManageLaptopListManager
+import com.example.jub_jub_android.entity.singleton.StudentRentStatusListManager
+import com.example.jub_jub_android.ui.adapter.recyclerview.MyRentList_RecyclerViewAdapter
 import kotlinx.android.synthetic.main.fragment_item_status_list.view.*
 
-class ManageLaptop_ViewpagerAdapter(var context: Context) : PagerAdapter() {
+class MyRentList_ViewPagerAdapter(context: Context) : PagerAdapter() {
 
+
+    private var context = context
     private lateinit var layoutInflater: LayoutInflater
-
 
     override fun getItemPosition(`object`: Any): Int {
         return POSITION_NONE
@@ -24,7 +25,7 @@ class ManageLaptop_ViewpagerAdapter(var context: Context) : PagerAdapter() {
     }
 
     override fun getCount(): Int {
-        return ManageLaptopListManager.getShowList().size
+        return StudentRentStatusListManager.getShowList().size
     }
 
     override fun destroyItem(container: ViewGroup, position: Int, `object`: Any) {
@@ -36,8 +37,8 @@ class ManageLaptop_ViewpagerAdapter(var context: Context) : PagerAdapter() {
         layoutInflater = context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
         val view: View = layoutInflater.inflate(R.layout.fragment_item_status_list, null)
 
-        //메인 화면 (기자재 목록)
-        var adapter = ManageLaptop_RecyclerViewAdapter(ManageLaptopListManager.getShowList()[position])
+        //나의 대여 목록.
+        var adapter = MyRentList_RecyclerViewAdapter(StudentRentStatusListManager.getShowList()[position])
         view.recyclerView_ItemStatusList.adapter = adapter
         adapter.notifyDataSetChanged()
